+++ src/components/ImageLoader/AnyLoader.jsx
@@ -49,7 +49,7 @@ export default (props) => {
     }
   }, [isLoaded, imgProps.src, placeholderProps.src, isComplete, onLoad]);
   useEffect(() => {
-    const placeholderCurrent = placeholder.props.blurhashRef?.current;
+    const placeholderCurrent = placeholder?.props.blurhashRef?.current;
     if (placeholderCurrent) {
       const computedStyle = getComputedStyle(ref.current);
       const { aspectRatio, objectFit } = computedStyle;